EllipticEnvelopeLearner: Move to outlier_detection module
